requirements.txt
requests
biopython
cachetools<4
click>=6.7
wrapt
pyyaml
decorator
bcbio-gff
pytest-timeit
